requirements.txt
customtkinter==5.2.2
python-docx==1.2.0
pillow==12.1.0
pywin32==311
packaging==26.0
pyinstaller